Separate logic in multiple files
